requirements.txt
altair==4.2.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
entrypoints==0.4
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
hydralit_components==1.0.10
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mlxtend==0.20.0
narwhals==1.9.4
numpy==1.23.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==3.20.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
Pympler==1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
scikit-learn==1.1.2
scipy==1.14.1
seaborn==0.13.2
semver==3.0.2
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.10.0
streamlit-lottie==0.0.3
streamlit-player==0.1.5
streamlit-tags==1.2.8
streamlit-vega-lite==0.1.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==5.0.3
zipp==3.20.2
